+++ CHANGELOG.md
@@ -143,6 +143,7 @@
 - Port the index controller to Symfony !3799 @\_fridtjof_
 - Remove the legacy entrypoint for routing and rendering pages !3799 @\_fridtjof_
 - Removed unnecessary database queries from the user details API endpoint !3807 @alex.simm
+- Removed unused script.js:initialize !3827 @chriswalg
 
 ## Dev/Test/CI stuff
 - Improved API tests runtime !3506 @alex.simm